+++ components/toggle-switch/src/define.ts
@@ -1,4 +1,4 @@
-import { ToggleSwitchElement } from './toggle-switch'
+import { ToggleSwitchElement } from './toggle-switch.js'
 
 if (!window.customElements.get(ToggleSwitchElement.defaultElementName)) {
     window.customElements.define(ToggleSwitchElement.defaultElementName, ToggleSwitchElement)